Apply a Filter to a Dataset or List of Datasets

Apply a Filter to a Dataset or List of Datasets

apply_filter(data, ...)

# S3 method for default
apply_filter(data, ...)

# S3 method for data.frame
apply_filter(data, id, target = deparse(substitute(data)), verbose = TRUE, ...)

# S3 method for list
apply_filter(data, id, verbose = TRUE, ...)

Arguments

data

data.frame of list of data.frames

...

Not used.

id

character. The ID of a filter defined with add_filter()

target

character. The name of the dataset, e.g. ADSL or ADTTE

verbose

logical. Should informative messages be printed? Defaults to TRUE.

Value

A new data.frame or list of data.frames filtered based upon the condition defined for id

Author

Thomas Neitmann (neitmant)

Examples

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae
datasets <- list(adsl = adsl, adae = adae)

add_filter("REL", "Related AEs", "ADAE", AEREL == "Y", overwrite = TRUE)

apply_filter(adsl, "SE")
#> Filter 'SE' matched target ADSL.
#> 400/400 records matched the filter condition `SAFFL == 'Y'`.
#> # A tibble: 400 × 55
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-C… id-128 CHN-3     32 YEARS M     ASIAN HISPA… CHN     Y    
#>  2 AB12345 AB12345-C… id-262 CHN-15    35 YEARS M     BLAC… NOT H… CHN     N    
#>  3 AB12345 AB12345-R… id-378 RUS-3     30 YEARS F     ASIAN NOT H… RUS     N    
#>  4 AB12345 AB12345-C… id-220 CHN-11    26 YEARS F     ASIAN NOT H… CHN     N    
#>  5 AB12345 AB12345-C… id-267 CHN-7     40 YEARS M     ASIAN NOT H… CHN     N    
#>  6 AB12345 AB12345-C… id-201 CHN-15    49 YEARS M     ASIAN NOT H… CHN     Y    
#>  7 AB12345 AB12345-U… id-45  USA-1     34 YEARS F     ASIAN NOT H… USA     N    
#>  8 AB12345 AB12345-U… id-261 USA-1     32 YEARS F     ASIAN NOT H… USA     N    
#>  9 AB12345 AB12345-N… id-173 NGA-11    24 YEARS F     BLAC… NOT H… NGA     N    
#> 10 AB12345 AB12345-C… id-307 CHN-1     24 YEARS M     ASIAN NOT H… CHN     Y    
#> # ℹ 390 more rows
#> # ℹ 44 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …
apply_filter(adae, "SER_REL")
#> Filters 'SER', 'REL' matched target ADAE.
#> 361/1934 records matched the filter condition `AESER == 'Y' & AEREL == 'Y'`.
#> # A tibble: 361 × 92
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-B… id-141 BRA-1     35 YEARS F     WHITE "NOT … BRA     N    
#>  2 AB12345 AB12345-B… id-236 BRA-1     32 YEARS M     BLAC… " NOT… BRA     N    
#>  3 AB12345 AB12345-B… id-236 BRA-1     32 YEARS M     BLAC… " NOT… BRA     N    
#>  4 AB12345 AB12345-B… id-236 BRA-1     32 YEARS M     BLAC… " NOT… BRA     N    
#>  5 AB12345 AB12345-B… id-265 BRA-1     25 YEARS M     WHITE "NOT … BRA     Y    
#>  6 AB12345 AB12345-B… id-42  BRA-1     36 YEARS M     BLAC… "NOT … BRA     N    
#>  7 AB12345 AB12345-B… id-42  BRA-1     36 YEARS M     BLAC… "NOT … BRA     N    
#>  8 AB12345 AB12345-B… id-42  BRA-1     36 YEARS M     BLAC… "NOT … BRA     N    
#>  9 AB12345 AB12345-B… id-65  BRA-1     25 YEARS F     BLAC… "NOT … BRA     N    
#> 10 AB12345 AB12345-B… id-93  BRA-1     34 YEARS F     ASIAN "HISP… BRA     Y    
#> # ℹ 351 more rows
#> # ℹ 81 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …
apply_filter(datasets, "SER_REL_SE")
#> Filter 'SE' matched target ADSL.
#> 400/400 records matched the filter condition `SAFFL == 'Y'`.
#> Filters 'SER', 'REL' matched target ADAE.
#> 361/1934 records matched the filter condition `AESER == 'Y' & AEREL == 'Y'`.
#> $adsl
#> # A tibble: 400 × 55
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-C… id-128 CHN-3     32 YEARS M     ASIAN HISPA… CHN     Y    
#>  2 AB12345 AB12345-C… id-262 CHN-15    35 YEARS M     BLAC… NOT H… CHN     N    
#>  3 AB12345 AB12345-R… id-378 RUS-3     30 YEARS F     ASIAN NOT H… RUS     N    
#>  4 AB12345 AB12345-C… id-220 CHN-11    26 YEARS F     ASIAN NOT H… CHN     N    
#>  5 AB12345 AB12345-C… id-267 CHN-7     40 YEARS M     ASIAN NOT H… CHN     N    
#>  6 AB12345 AB12345-C… id-201 CHN-15    49 YEARS M     ASIAN NOT H… CHN     Y    
#>  7 AB12345 AB12345-U… id-45  USA-1     34 YEARS F     ASIAN NOT H… USA     N    
#>  8 AB12345 AB12345-U… id-261 USA-1     32 YEARS F     ASIAN NOT H… USA     N    
#>  9 AB12345 AB12345-N… id-173 NGA-11    24 YEARS F     BLAC… NOT H… NGA     N    
#> 10 AB12345 AB12345-C… id-307 CHN-1     24 YEARS M     ASIAN NOT H… CHN     Y    
#> # ℹ 390 more rows
#> # ℹ 44 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …
#> 
#> $adae
#> # A tibble: 361 × 92
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-B… id-141 BRA-1     35 YEARS F     WHITE "NOT … BRA     N    
#>  2 AB12345 AB12345-B… id-236 BRA-1     32 YEARS M     BLAC… " NOT… BRA     N    
#>  3 AB12345 AB12345-B… id-236 BRA-1     32 YEARS M     BLAC… " NOT… BRA     N    
#>  4 AB12345 AB12345-B… id-236 BRA-1     32 YEARS M     BLAC… " NOT… BRA     N    
#>  5 AB12345 AB12345-B… id-265 BRA-1     25 YEARS M     WHITE "NOT … BRA     Y    
#>  6 AB12345 AB12345-B… id-42  BRA-1     36 YEARS M     BLAC… "NOT … BRA     N    
#>  7 AB12345 AB12345-B… id-42  BRA-1     36 YEARS M     BLAC… "NOT … BRA     N    
#>  8 AB12345 AB12345-B… id-42  BRA-1     36 YEARS M     BLAC… "NOT … BRA     N    
#>  9 AB12345 AB12345-B… id-65  BRA-1     25 YEARS F     BLAC… "NOT … BRA     N    
#> 10 AB12345 AB12345-B… id-93  BRA-1     34 YEARS F     ASIAN "HISP… BRA     Y    
#> # ℹ 351 more rows
#> # ℹ 81 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …
#>