Create a score object for ANOVA F-test F-statistics and p-values

Construct a score object containing metadata for univariate feature scoring using the ANOVA F-test. Output a score object containing associated metadata such as range, fallback_value, score_type ("fstat" or "pval"), direction, and other relevant attributes.

score_aov(
  range = c(0, Inf),
  fallback_value = Inf,
  score_type = "fstat",
  direction = "maximize"
)

Arguments

range

A numeric vector of length two, specifying the minimum and maximum possible values, respectively.

fallback_value

A numeric scalar used as a fallback value. Typical values include:

0

Inf (default)

For F-statistics, the fallback_value should be "Inf". For p-values, since the default applies a negative log10 transformation to p-values, the fallback_value should be "Inf".

score_type

A character string indicating the type of scoring metric to compute. Available options include:

"fstat"

"pval"

direction

A character string indicating the optimization direction. One of:

"maximize" (default)

"minimize"

"target"

For F-statistics, the direction should be "maximize". For p-values, since the default applies a negative log10 transformation to p-values, the direction should be "maximize".

Value

A score object containing associated metadata such as range, fallback_value, score_type ("fstat" or "pval"), direction, and other relevant attributes.

Examples

# Create a score object
score_aov()
#> ANOVA F-test F-statistics and p-values
# Change score type
score_obj <- score_aov()
score_obj$score_type <- "pval"